Chore: move theme CodeMirror extension into /src/editor/plugins/theme
Moves this extension that was in a stray file called
`/src/lib/codeEditor.ts` to make this CodeMirror extension follow the
convention of our others. Broken out from
https://github.com/KittyCAD/modeling-app/pull/9516 because it is just
diff noise.
